virt-resize: Take into account large start offset of the first partition when calculating overhead (RHBZ#974904).
Since we don't usually move the first partition, if the first
partition has an unusually large offset from the start of the disk,
then the unpartitioned space in front of that partition counts as
partitioning overhead.  However the previous surplus calculation was
not taking that into account.

This was a problem for certain Ubuntu images which are partitioned
with an 8 MB gap before the first partition.

Thanks: David Hart.
(cherry picked from commit b730bc0c4685887232daed8fafa491b28f38c300)